+++ functions/modules/openfaas/service-executor/src/main/java/org/acme/quickstart/resources/ResourcesLocator.java
@@ -1,7 +1,9 @@
 package org.acme.quickstart.resources;
 
+import java.math.BigDecimal;
+
 public interface ResourcesLocator {
 
-    int getUsedCpuPercentage();
+    BigDecimal getUsedCpuPercentage();
     
 }
